requirements.txt
Adafruit-ADS1x15==1.0.2
Adafruit-Blinka==6.11.1
adafruit-circuitpython-lis3dh==5.1.10
Adafruit-DHT==1.4.0
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.15.3
Adafruit-PureIO==1.1.9
APScheduler==3.7.0
cffi==1.14.6
click==8.0.1
cryptography==3.4.7
cycler==0.10.0
Flask==2.0.1
Flask-APScheduler==1.12.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.0
pandas==1.3.0
Pillow==8.3.1
pkg-resources==0.0.0
pycparser==2.20
pyftdi==0.53.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.2
pytz==2021.1
pyusb==1.1.1
rpi==0.0.1
rpi-ws281x==4.3.0
RPi.GPIO==0.7.0
six==1.16.0
smbus==1.1.post2
spidev==3.5
sysv-ipc==1.1.0
typing-extensions==3.10.0.0
tzlocal==2.1
Werkzeug==2.0.1
zipp==3.5.0
